add notion of global/local KnowledgeSources
